Move tunable to rocm ep level
