+++ onnxruntime/contrib_ops/rocm/bert/util.h
@@ -27,20 +27,6 @@ struct alignas(sizeof(T) * VecSize) AlignedVector {
   T val[VecSize];
 };
 
-class Timer {
- public:
-  explicit Timer(hipStream_t stream);
-  void Start();
-  void End();
-  float Duration();
-  ~Timer();
-
- private:
-  hipStream_t stream_;
-  hipEvent_t start_;
-  hipEvent_t end_;
-};
-
 }  // namespace rocm
 }  // namespace contrib
 }  // namespace onnxruntime
